// src/store/modules/Setting/types.ts
export default interface settingStateTypes {
  title: string
  fixedHeader: Boolean
  sideBarLogo: Boolean
  showSettings: Boolean
  tableHeight: Number
  hideHeader: boolean
  lang: string
  tagSlider: number
  annoLayout: Array<any>
  reviewLayout: Array<any>
  oldVersion: Boolean
  isDraggable: Boolean
  tagHorizontalLayout: Boolean
}